requirements.txt
dnspython
ipaddr
pygeoip
Queue
